% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chloride.R
\docType{data}
\name{chloride}
\alias{chloride}
\title{Effect of ammonium chloride on yield}
\format{
A data frame with 32 rows and 5 columns:
\describe{
  \item{block}{Experimental block}
  \item{unit}{Production unit}
  \item{amount}{Ammonium chloride amount, `low` or `high`}
  \item{quality}{Ammonium chloride quality, `low` or `high`}
  \item{yield}{Process yield in pounds}
}
}
\source{
{Davies, O.L. (ed.) (1956) \emph{The Design and Analysis of
  Industrial Experiments}, 2nd Edition, London: Longman, 375.}
}
\usage{
chloride
}
\description{
Data from an experiment to investigate the yield of an organic chemical from
a process which used ammonium chloride.
}
\details{
The main interest lay in the comparison between the use of finely ground and
coarse ammonium chloride; the fine grade cost more. The experimenters were
also interested in the effect of adding an extra 10% of ammonium chloride
above the usual amount. There were two production units at the plant, and it
was thought that the unit used would have an effect on yield, but that
interactions between the unit and the two factors involving ammonium chloride
were unlikely. The experiment thus had a 23 factorial treatment structure.
Experimental runs were carried out in blocks, each block using a blend of
input materials, and it was not desirable for these blends to be too large.
Therefore a complete randomized block design was not considered appropriate;
instead, blocks of size 4 were used and the 3-way interaction was confounded
with blocks.
}
\examples{
library(hsds)
library(ggplot2)

# Two-way interaction plot between amount and quality
chloride \%>\%
  dplyr::filter(!is.na(yield)) \%>\%
  ggplot(aes(x = amount, y = yield, color = quality, group = quality)) +
  geom_point() +
  geom_line() +
  facet_wrap(
    .~block,
    nrow = 2,
    labeller = labeller(.rows = ~paste0("Block ", .x))
  ) +
  theme_classic()
}
\keyword{datasets}
