% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{data_list}
\alias{data_list}
\title{List the data sets already processed in the package}
\usage{
data_list(print = 5)
}
\arguments{
\item{print}{Number of unprocessed data sets to print (default is 5)}
}
\description{
List the data sets already processed in the package
}
\keyword{internal}
