% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{find_data}
\alias{find_data}
\title{Find a data set for a specific topic}
\usage{
find_data(topic, max_print = 10)
}
\arguments{
\item{topic}{string. Topic of interest in the package}

\item{max_print}{numeric. Maximum number of data sets to print if more are
found in the package}
}
\description{
Find all the data sets in the package with a keyword of interest. For each
data set, the following information are printed:
\itemize{
\item Name of the data set (as a link to the reference page online !)
\item Size (rows X columns)
\item One-line description
}
}
\examples{
find_data(topic = "non-linear")

}
