% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\docType{data}
\name{interval}
\alias{interval}
\title{Interval between cars on the M1 motorway}
\format{
A data frame with 48 rows and 3 columns:
\describe{
\item{car}{Unique ID to define the cars}
\item{time}{Time at which the car passed the junction. Encoded as a string in the 'HH:MM:SS' format}
}
}
\source{
{Lewis, T. and the M345 Course Team (1986) \emph{M345 Statistical Methods, Unit 2: Basic Methods: Testing and Estimation}, Milton Keynes: The Open University, 16.}
}
\usage{
interval
}
\description{
Times that 41 successive vehicles travelling northwards along the M1 motorway
in England passed a fixed point near Junction 13 in Bedfordshire on Saturday
23 March 1985.
The times are recorded on the 24-hour clock to the nearest second.
}
\details{
The data were collected by Professor Toby Lewis as a sample of event times in
a point process which it may be appropriate to model as a Poisson process.
They are also ideal to learn to handle time objects in R.
}
\keyword{datasets}
