% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths.R
\docType{data}
\name{lengths}
\alias{lengths}
\title{Guessing lengths}
\format{
A data frame with 113 rows and 3 columns:
\describe{
\item{unit}{Unit used for guessing (feet or metres)}
\item{guessed_length}{Length guessed by the students}
\item{true_length}{True length of the room, in corresponding units}
}
}
\source{
{Hills, M. and the M345 Course Team (1986) \emph{M345 Statistical Methods, Unit 1: Data, distributions and uncertainty}, Milton Keynes: The Open University. Tables 2.1 and 2.4.}
}
\usage{
lengths
}
\description{
Students guessing the length of a room in metres and in feet.
}
\details{
Shortly after metric units of length were officially introduced in Australia,
each of a group of 44 students was asked to guess, to the nearest metre, the
width of the lecture hall in which they were sitting.
Another group of 69 students in the same room was asked to guess the width in
feet, to the nearest foot. The true width of the hall was 13.1 metres (43.0
feet).
}
\keyword{datasets}
