% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastes.R
\docType{data}
\name{pastes}
\alias{pastes}
\title{Strength of chemical pastes}
\format{
A data frame with 60 rows and 4 columns:
\describe{
\item{batch}{Factor. Batch number (1 to 10)}
\item{cask}{Factor. Cask number within each cask(1 to 3)}
\item{sample}{Factor. Sample number within each cask (1 or 2)}
\item{strength}{Percentage paste strength in the analysed sample}
}
}
\source{
Davies, O.L. and Goldsmith, P.L. (eds.) (1972) \emph{Statistical
Methods in Research and Production}, 4th Edition, Edinburgh: Oliver and
Boyd, 136.
}
\usage{
pastes
}
\description{
A chemical paste product is delivered in casks, and there are
thought to be variations in the mean strength of the paste between delivery
batches. When a batch arrives, the paste is tested by sampling casks from
the batch and then analysing material from the casks. Further sources of
error arise in the sampling and in the analysis. The data come from ten
randomly chosen delivery batches; from each, three casks are selected at
random, and two analyses are carried out on the contents of each selected
cask. The response is the percentage paste strength in the analysed sample.
}
\details{
Davies and Goldsmith use the following data to demonstrate the use
of analysis of variance in a hierarchical situation where the aim is to
estimate variance components.

This data is also useful to demonstrate the difference between crossed and
nested random effect structures in linear mixed models in R using the
\code{lme4} package.
}
\examples{
library(lme4)

# The cask numbers are not uniquely defined across batches
xtabs(~ batch + cask, pastes)

# This can lead to a "crossed" random effect design if the model terms are
# not correctly specified
m1 <- lmer(strength ~ (1 | batch) + (1 | cask), data = pastes)
summary(m1)
# The model identifies 3 different casks used repeatedly over the 10 batches (WRONG)
summary(m1)$ngrps

# Specifying the random effect structure avoids this issue, and correctly leads
# to a "nested" random effect structure
m2 <- lmer(strength ~ (1 | batch / cask), data = pastes)
summary(m2)
# There are indeed 30 combinations of casks and batches (3 *different* casks used
# in each batch)
summary(m2)$ngrps
}
\keyword{datasets}
