% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piston.R
\docType{data}
\name{piston}
\alias{piston}
\title{Piston-ring failures}
\format{
A data frame with X rows and X columns:
\describe{
  \item{compressor}{Compressor ID. Numeric factor.}
  \item{leg}{Compressor leg. Factor with 3 levels `North`, `Center`, `South`.}
  \item{failures}{Number of failures}
}
}
\source{
{Davies, O.L. and Goldsmith, P.L. (eds.) (1972) \emph{Statistical
  Methods in Research and Production}, 4th Edition, Edinburgh: Oliverand
  Boyd, 324.}
}
\usage{
piston
}
\description{
Data on the number of failures of piston-rings in each of three legs in each
of four steam-driven compressors located in the same building. The
compressors have identical design and are oriented in the same way. Questions
of interest are whether the probability of failure varies between compressors
or between different legs, and whether the pattern of the location of
failures is different for different compressors. Davies and Goldsmith
investigate these questions using chi-squared tests; log-linear analysis
might be an alternative.
}
\examples{
library(ggplot2)
library(hsds)

# Plot of the number of failures per leg, per compressor
ggplot(piston, aes(x = leg, y = failures)) +
  geom_boxplot() +
  geom_point(aes(color = compressor), size = 2) +
  theme_classic(12) +
  ggeasy::easy_labs()

# Questions of interest:
#   - Does the probability of failure vary between compressors or between
#     different legs?
#   - Is the pattern of the location of failures is different for different
#     compressors?

# Davies and Goldsmith investigate these questions using chi-squared tests
tab <- xtabs(failures ~ compressor + leg, data = piston)
chisq.test(tab)
}
\keyword{datasets}
