% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{print_dataset_info}
\alias{print_dataset_info}
\title{Print information about a data set using the `cli` package}
\usage{
print_dataset_info(name, description, size, ...)
}
\description{
Print information about a data set using the `cli` package
}
\keyword{internal}
