% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/software.R
\docType{data}
\name{software}
\alias{software}
\title{Software system failures}
\format{
A data frame with X rows and X columns:
\describe{
  \item{time}{Time to failure in CPU seconds}
}
}
\source{
{Musa, J.D., Iannino, A. and Okumoto, K. (1987) \emph{Software reliability: measurement,prediction, application}, New York: McGraw-Hill, 305.}
}
\usage{
software
}
\description{
These data give the failure times (in CPU seconds, measured in terms of
execution time) of a real-time command and control software system.
The data can be used as an example for fitting various reliability models.
Musa et al. recommend fitting a nonhomogenous Poisson process with linearly
exponentially decreasing failure time.
}
\examples{
# Load necessary libraries
library(hsds)

# Load the software dataset
data("software")

# Distribution of the failure time
with(
  software,
  hist(
    time,
    main = "Distribution of the failure times",
    xlab = "Failure time (CPU seconds)"
  )
)

# Plot of the cumulative number of events over time
with(
  software,
  plot(
    time,
    failure_id,
    main = "Cumulative number of failures over time",
    xlab = "Time (CPU seconds)", ylab = "Number of failures"
  )
)
}
\keyword{datasets}
