% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.R
\docType{data}
\name{water}
\alias{water}
\title{Mortality and water hardness}
\format{
A data frame with 61 rows and 5 columns:
\describe{
  \item{town}{Name of the town, as a factor variable}
  \item{mortality}{Mortality rate per 100,000 inhabitants}
  \item{calcium}{Calcium concentration in the drinking water in parts per million (ppm)}
  \item{lon}{Longitude using the EPSG:4326 coordinate system}
  \item{lat}{Latitude using the EPSG:4326 coordinate system}
}
}
\source{
{Hills, M. and the M345 Course Team (1986) \emph{M345 Statistical Methods, Unit 3: Examining Straight-line Data}, Milton Keynes: The Open University, 28. Data provided by Professor M.J. Gardner, Medical Research Council Environmental Epiderniology Research Unit, Southampton.}
}
\usage{
water
}
\description{
Annual mortality rate per 100 000 for males, averaged over the years 1958-1964, and the calcium concentration (in parts per million) in the drinking water supply for 61 large towns in England and Wales.
}
\examples{
library(sf)
library(ggplot2)
library(rnaturalearth)
library(hsds)

# Retrieve base map of England
england <- ne_states(country = "united kingdom")

# Convert long and lat into `sf` geometry
water \%>\%
  st_as_sf(coords = c("lon", "lat"), crs = 4326) \%>\%
  ggplot() +
  geom_sf(data = england, fill = NA) +
  geom_sf(aes(fill = mortality), shape = 21, size = 2) +
  coord_sf(xlim = c(-10, 2), ylim = c(50, 56)) +
  scale_fill_gradient(
    name = "Mortality per 100,000 hab.",
    low = "lightgrey",
    high = "darkred"
  ) +
  theme_bw(13) +
  theme(legend.position = "bottom")
}
\keyword{datasets}
