% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\docType{data}
\name{weight}
\alias{weight}
\title{Weight gain in rats}
\format{
A data frame with 30 rows and 3 columns:
\describe{
  \item{weight}{Weight gain in grams}
  \item{source}{Protein source (`beef` or `cereal`)}
  \item{amount}{Protein amount (`low` or `high`)}
}
}
\source{
{Snedecor, G.W. and Cochran, G.C. (1967) \emph{Statistical Methods},
  6th edition, Ames, Iowa: Iowa State University Press, 347.}
}
\usage{
weight
}
\description{
The data come from an experiment to study the gain in weight of rats fed on
four different diets, distinguished by amount of protein (low and high) and
by source of protein (beef and cereal). The design of the experiment is
completely randomized with ten rats on each of the four treatments (which
have a complete factorial structure).
}
\examples{
library(hsds)
library(ggplot2)

# Mean weight gain per diet
weight \%>\%
  dplyr::summarise(
    mean(weight),
    sd(weight),
    .by = c(source, amount)
  )

# Colored box plot
ggplot(weight, aes(x = amount, y = weight, color = source)) +
  geom_boxplot() +
  geom_jitter(position = position_dodge(width = 0.75)) +
  theme_classic()

# Two-way ANOVA
aov(weight ~ amount * source, data = weight) \%>\%
  summary()
}
\keyword{datasets}
