% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cement.R
\docType{data}
\name{cement}
\alias{cement}
\title{Tensile strength of cement}
\format{
A data frame with 21 rows and 2 columns:
\describe{
  \item{time}{Curing time in days}
  \item{strength}{Tensile strength (kg/cm2)}
}
}
\source{
{Hald, A. (1952) \emph{Statistical Theory with Engineering
  Applications}, New York: John Wiley & Sons, 451.}
}
\usage{
cement
}
\description{
Tensile strength of different batches of cement after different curing times.
The tensile strength of cement depends on (among other things) the length of
time for which the cement is dried or 'cured'. The relationship between
curing time and strength is non-linear. Hald regresses log tensile strength
on the reciprocal of curing time.
}
\examples{
library(ggplot2)
library(hsds)

# There is a clear non-linear relation between time and strength
# Hald regresses log tensile strength on the reciprocal of curing time.

ggplot(cement, aes(y = strength, x = time)) +
  geom_point() +
  geom_smooth(method = "lm", formula = y ~ exp(1 / x), se = FALSE) +
  theme_classic(base_size = 12)

# Using Hald's method the relationship is clearly linear
ggplot(cement, aes(y = log(strength), x = 1 / time)) +
  geom_point() +
  geom_smooth(method = "lm", formula = y ~ x, se = FALSE) +
  theme_classic(base_size = 12)

# A linear model indeed fits well (high R-squared) and
# the reciprocal of time is significant
lm(log(strength) ~ I(1 / time), data = cement) |>
  summary()
}
\keyword{Non-linear}
