% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{data_setup}
\alias{data_setup}
\title{Setup for a new data set}
\usage{
data_setup(data)
}
\arguments{
\item{data}{character. Name of the data set for which the files are created. Must be an existing \code{.dat} file in the \code{data-raw} folder.}
}
\description{
Generate the three files needed for a new data set:
\itemize{
\item A file to import the raw data as an R data set
\item A file to generate the documentation of the data set
\item A file to contain the examples for the data generated
}
}
\keyword{internal}
