% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piston.R
\docType{data}
\name{piston}
\alias{piston}
\title{Piston-ring failures}
\format{
A data frame with X rows and X columns:
\describe{
  \item{comp_id}{Compressor ID (1 to 4)}
  \item{leg}{Compressor leg (North, Center, South)}
  \item{n_failures}{Number of failures observed)}
}
}
\source{
Davies, O.L. and Goldsmith, P.L. (eds.) (1972) \emph{Statistical Methods in Research and Production}, 4th Edition, Edinburgh: Oliverand Boyd, 324.
}
\usage{
piston
}
\description{
Number of failures of piston-rings in each
of three legs in each of four steam-driven compressors located in the same
building. The compressors have identical design and are oriented in the same
way.
}
\details{
Questions of interest are whether the probability of failure varies
between compressors or between different legs, and whether the pattern of the
location of failures is different for different compressors. Davies and
Goldsmith investigate these questions using chi-squared tests; log-linear
analysis might be an alternative.
}
\examples{
tab <- xtabs(n_failures ~ comp_id + leg, data = piston)
tab

chisq.test(tab)
}
\keyword{datasets}
